requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
pydantic==1.10.12
langchain==0.0.220
openai==0.28.1
faiss-cpu==1.7.4
numpy==1.25.2
motor==3.2.0
python-dotenv==1.0.0
pyjwt==2.8.0
sentry-sdk==1.28.1
pytest==7.4.0
PyYAML==6.0
